API

API

Whale.CCD — Type.

CCD{<:Real,RecTree}

Conditional clade distribution with many helper fields. See read_ale for details on IO.

julia> x
CCD{Float64,PhyloTrees.RecTree}(13 taxa, 83 clades, 5001 samples)

julia> x.ccp
Dict{Tuple,Float64} with 203 entries:
  (83, 65, 66) => 0.014797
  (46, 6, 60)  => 0.0185963
  (25, 16, 23) => 1.0
  (43, 7, 38)  => 0.995601
  ⋮            => ⋮

julia> Whale.get_triples(x, 68)
3-element Array{Tuple{Int64,Int64,Int64},1}:
 (24, 75, 1)
 (2, 74, 22)
 (6, 36, 103)

source

Whale.GBMModel — Type.

GBMModel(st::SlicedTree, ν::T, η::T, λ::T, μ::T, q::T) where
    T::Union{<:Distribution,Array{<:Distribution,1},<:Real}

Hierarchical model for Whale using a GBM prior.

ν: autocorrelation strength (variance of geometric Brownian motion)

η: parameter of geometric prior on the number of genes at the root of the species tree S

λ: duplication rate at the root of S

μ: loss rate at the root of S

q: retention rates

Example: InverseGamma prior on ν, Beta prior on η, Exponential priors on λ and μ at the root and Beta(1,1) priors on the retention rates q.

julia> m = GBMModel(st, InverseGamma(15), Beta(10, 1), Exponential(1), Exponential(1), [Beta(1,1) for i=1:nwgd(st)]);

julia> rand(m, st)  # sample a state from the prior
Dict{Symbol,Union{Float64, Array{Float64,1}}} with 5 entries:
  :ν => 0.0771695
  :μ => [0.498764, 0.500195, 0.459161, 0.615959, 0.482646, 0.502002, 0.402976, 0.354524, 0.378785, …
  :λ => [0.760383, 0.733029, 0.696493, 0.592979, 0.718899, 1.01307, 0.642244, 0.639109, 0.600577, 0…
  :η => 0.915222
  :q => [0.761553, 0.729741, 0.116805, 0.278579, 0.200221, 0.293774, 0.869646]

source

Whale.IRModel — Type.

IRModel(st::SlicedTree, ν::T, η::T, λ::T, μ::T, q::T) where
    T::Union{<:Distribution,Array{<:Distribution,1},<:Real}

Hierarchical model for Whale using a independent rates prior. As GBMModel but ν is the variance of the log-Normal prior on the branch-wise rates.

Example:

julia> m = IRModel(st, InverseGamma(10), Beta(6,3));

julia> m.ν
InverseGamma{Float64}(
invd: Gamma{Float64}(α=10.0, θ=1.0)
θ: 1.0
)

julia> m.λ
Exponential{Float64}(θ=1.0)

julia> rand(m, st)
Dict{Symbol,Union{Float64, Array{Float64,1}}} with 5 entries:
  :ν => 0.0955299
  :μ => [0.456988, 0.43639, 0.626645, 0.464031, 0.540543, 0.629841, 0.574996, 0.498506, 0.428841, 0…
  :λ => [0.498921, 0.69414, 0.54769, 0.533595, 0.521085, 0.515748, 0.534503, 0.594045, 0.583829, 0.…
  :η => 0.486934
  :q => [0.970779, 0.965934, 0.788127, 0.405944, 0.827755, 0.253859, 0.395734]

source

Whale.SlicedTree — Type.

SlicedTree(tree::Arboreal, wgdconf=Dict(), Δt=0.05, minn=5, maxn=Inf)

Sliced species tree with optional WGD nodes. WGDs are specified using a configuration dict such as for example:

wgdconf = Dict(
    "PPAT" => ("PPAT", 0.655),
    "CPAP" => ("CPAP", 0.275),
    "BETA" => ("ATHA", 0.55),
    "ANGI" => ("ATRI,ATHA", 3.08),
    "SEED" => ("GBIL,ATHA", 3.9),
    "MONO" => ("OSAT", 0.91),
    "ALPH" => ("ATHA", 0.501))

where the keys are WGD IDs (names), and values are given as a tuple or array with as first entry a single taxon or pair of taxa (comma-separated) that specifies the last common ancestor node that is preceded by the WGD of interest and the second entry the estimate time (before present) of the WGD event. Δt is the desired slice length (time interval), minn the minimum number of slices a branch should have, and maxn the maximum number of slices a branch should have.

source

Whale.WhaleChain — Type.

WhaleChain(st::SlicedTree, π::Model)

Chain object for performing MCMC under various hierarchical models defined in Whale.

julia> w = WhaleChain(st, IRModel(st))
WhaleChain{IRModel}(SlicedTree(9, 17, 7))

julia> w[:q]  # current state's q values
7-element Array{Float64,1}:
 0.4068351791361286
 0.5893350551453437
 0.21472854312451684
 0.8068526003250731
 0.05239284527812649
 0.8432325466244709
 0.9006557436550706

julia> w[:μ, 3]  # current state of μ for branch 3
0.6578486200316909

julia> logpdf(w)  # prior logpdf
59.40417835345352

julia> logpdf(w, :ν=>0.1)  # prior logpdf, but with ν = 0.1
58.49660562613573

julia> logpdf(w, :η=>0.2)  # prior logpdf, but with η = 0.2
45.22412219063766

julia> WhaleModel(w)  # WhaleModel corresponding to current state of the chain
WhaleModel{Float64,CCD}(
λ: [0.440648, 0.486472, 0.412741, 0.432659, 0.392268, 0.485483, 0.485181, 0.422687, 0.431686, 0.477667, 0.405392, 0.414238, 0.442341, 0.390893, 0.429505, 0.41379, 0.435038]
μ: [0.635544, 0.575146, 0.657849, 0.657189, 0.626361, 0.668117, 0.619245, 0.600366, 0.682475, 0.570257, 0.58954, 0.576526, 0.6966, 0.729826, 0.561087, 0.614903, 0.668041]
q: [0.406835, 0.589335, 0.214729, 0.806853, 0.0523928, 0.843233, 0.900656]
η: 0.966691254252368

source

Whale.WhaleModel — Type.

WhaleModel{T<:Real,CCD}(S::SlicedTree, λ, μ, q, η=0.9, cond="oib")

The Whale probabilistic model, containing both the sliced species tree, parameters and fields for extinction and propagation probabilities.

julia> st = Whale.example_tree();

julia> ccd = read_ale("example/example-ale/", st);
[ Info:  .. read 12 ALE files

julia> w = WhaleModel(st, 0.2, 0.3, 0.1, 0.9)
WhaleModel{Float64,CCD}(
λ: [0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2]
μ: [0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3, 0.3]
q: [0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1]
η: 0.9
)

julia> logpdf(w, ccd)
-296.67809475568924

julia> ccd[1].tmpmat  # this did not store the dynamic programming (DP) matrix
Dict{Int64,Array{Float64,2}} with 0 entries

julia> logpdf(w, ccd, matrix=true)  # this will store the DP matrix
-296.67809475568924

julia> ccd[1].tmpmat  # this is the DP matrix, required for backtracking
Dict{Int64,Array{Float64,2}} with 24 entries:
  1  => [0.000160506; 0.00276326; … ; 6.63525e-24; 4.18032e-15]
  2  => [0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0]
  16 => [0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0]
  ...

source

Whale.backtrack! — Function.

backtrack!(ccd::CCD, w::WhaleModel, n=1)
backtrack!(D::CCDArray, w::WhaleModel, n=1)
backtrack!(D::CCDArray, w::WhaleChain, n=1000)
backtrack!(D::CCDArray, df::DataFrame, st::SlicedTree, n=1000)

Sample reconciled trees (RecTree) from the Whale model conditional on the conditional clade distribution (CCD) by stochastic backtracking along the dynamic programming matrix as in Szollosi 2013. If the recmat field of the ccd argument is non-empty, this matrix will be used, otherwise the reconciliation matrix is computed from scratch. Reconciled trees are stored in the CCD object for convenience.

Note

If the second argument is a WhaleChain object (or data frame), than post hoc backtracking is performed, that is, reconciled trees are simulated from the posterior predictive distribution after running an MCMC chain.

source

Whale.consensus — Function.

consensus(x::CCD, S::SlicedTree)
consensus(D::DArray, S::SlicedTree)

Get consensus trees from the backtracked RecTrees in the rectrs field of the conditional clade distribution object(s) (CCD).

Warning

Branch lengths are not yet correct (they are meaningfully related to the true branch length estimates nevertheless).

Warning

A consensus reconciled tree does not include loss events. This is because different reconciled trees for the same family can have different numbers of loss events.

source

Whale.contreetable — Method.

contreetable([io::IO,] contree::ConRecTree, S::SlicedTree)

Write a table representation of a consensus reconciled tree.

source

Whale.describe — Method.

describe(w::WhaleModel)

Get a detailed description of a particular WhaleModel.

source

Whale.mcmc! — Method.

mcmc!(w::WhaleChain, D::CCDArray, n::Int64, args...; kwargs...)

Perform n generations of MCMC sampling for a WhaleChain given a bunch of observed CCDs.

source

Whale.mle — Function.

mle(w::WhaleModel, ccd::CCDArray, optimizer=LBFGS(); kwargs...)

source

Whale.read_ale — Method.

read_ale(fname::String, s::SlicedTree)

Read in a bunch of conditional clade distributions (CCD) from ALEobserve (.ale) files. Either provide

a file with a filename on each line

a directory with .ale files

a single .ale file

an empty file, for running MCMC under the prior alone

julia> st = Whale.example_tree()
SlicedTree(9, 17, 7)

julia> ccd = read_ale("example/example-ale/", st)
[ Info:  .. read 12 ALE files
12-element DistributedArrays.DArray{CCD,1,Array{CCD,1}}:
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 83 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 55 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 89 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 131 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 107 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 59 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 53 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 83 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 59 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 95 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 67 clades, 5001 samples)
 CCD{Float64,PhyloTrees.RecTree}(13 taxa, 65 clades, 5001 samples)

source

Whale.GeometricBrownianMotion — Type.

GeometricBrownianMotion{T<:Real}
GBM(t::SlicedTree, r::T, v::T) where {T<:Real}

Distribution induced by a Geometric Brownian Motion (GBM) over the SlicedTree. r is the value at the root of the tree, ν is the variance (autocorrelation strength).

The log density for the GBM distribution is computed with an implementation of the GBM prior on rates based on Ziheng Yang's MCMCTree, described in Rannala & Yang (2007). This uses the approach whereby rates are defined for midpoints of branches, and where a correction is performed to ensure that the correlation is proper (in contrast with Thorne et al. 1998).

source